+++ .gitignore
@@ -20,5 +20,4 @@
 bin/
 release/
 coverage/
-coverage.*
-
+coverage*.**
