+++ Makefile
@@ -40,13 +40,14 @@ build: tidy check
 
 unit_test:
 	@echo "run unit test"
-	@go test -race -coverprofile=coverage.txt -covermode=atomic -v ./...
-	@go tool cover -html="coverage.txt" -o "coverage.html"
+	@go test -race -cover -coverprofile=coverage_unit.txt -v ./...
+	@go tool cover -html="coverage_unit.txt" -o "coverage_unit.html"
 	@echo "ok"
 
 integration_test:
 	@echo "run integration test"
-	@go test -race -tags tests -v ./...
+	@go test -race -tags tests -cover -coverprofile=coverage_integration.txt -v ./...
+	@go tool cover -html="coverage_integration.txt" -o "coverage_integration.html"
 	@echo "ok"
 
 tidy:
